# specs/001-tharwa-platform-mvp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Tharwa API
  version: 1.0.0
  description: >
    Backend for Tharwa (ثروة). Public read endpoints are unauthenticated for MVP market browsing.
    Admin routes are under /admin/v1 (separate security scheme). Parallel EGP is not in MVP.
servers:
  - url: https://api.example.com
    description: Production (replace with real host)
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /v1/stocks:
    get:
      tags: [Market]
      summary: Curated Egyptian equities list
      operationId: listStocks
      responses:
        "200":
          description: Curated list
          content:
            application/json:
              schema:
                type: object
                required: [items, disclaimer]
                properties:
                  disclaimer:
                    type: string
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/EquityListItem"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/stocks/{symbol}:
    get:
      tags: [Market]
      summary: Equity detail headline fields
      operationId: getStockDetail
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
            example: EGX:COMI
      responses:
        "200":
          description: Detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquityDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/stocks/{symbol}/history:
    get:
      tags: [Market]
      summary: Historical candles for charting
      operationId: getStockHistory
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: range
          in: query
          schema:
            type: string
            enum: [1d, 1w, 1m, 1y]
            default: 1m
      responses:
        "200":
          description: OHLCV series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistorySeries"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /v1/journal:
    get:
      tags: [Consumer]
      summary: List trade journal entries
      operationId: listJournal
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Journal lines
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Consumer]
      summary: Add journal entry
      operationId: createJournalEntry
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [item]
                properties:
                  item:
                    $ref: "#/components/schemas/JournalEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/journal/{id}:
    patch:
      tags: [Consumer]
      summary: Update journal entry (note, price, executed date)
      operationId: patchJournalEntry
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalPatchRequest"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                type: object
                required: [item]
                properties:
                  item:
                    $ref: "#/components/schemas/JournalEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Consumer]
      summary: Delete journal entry
      operationId: deleteJournalEntry
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/portfolio/summary:
    get:
      tags: [Consumer]
      summary: Portfolio summary from journal (self-reported)
      operationId: getPortfolioSummary
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Aggregated positions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/portfolio/positions/{instrumentId}:
    delete:
      tags: [Consumer]
      summary: Remove position (deletes all journal lines for instrument)
      operationId: deletePortfolioPosition
      security:
        - BearerAuth: []
      parameters:
        - name: instrumentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Position removed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    BadRequest:
      description: Validation or bad input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Unauthorized:
      description: Missing or invalid credentials / token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
...
    ServiceUnavailable:
      description: Upstream unavailable and no acceptable cache
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    QuoteMeta:
      type: object
      required: [asOf, quoteCategory, sessionState, isStale]
      properties:
        asOf:
          type: string
          format: date-time
        quoteCategory:
          type: string
          enum: [official, indicative, estimate]
        sessionState:
          type: string
          enum: [open, closed, pre, post, unknown]
        isStale:
          type: boolean
...
    EquityListItem:
      allOf:
        - $ref: "#/components/schemas/QuoteMeta"
        - type: object
          required: [instrumentId, symbol, name]
          properties:
            instrumentId:
              type: string
              format: uuid
            symbol:
              type: string
            name:
              type: string
            nameAr:
              type: string
              description: Official display name in Arabic (curated instruments)
            last:
              type: number
              format: double
              nullable: true
            changePct:
              type: number
              format: double
              nullable: true
    EquityDetail:
      allOf:
        - $ref: "#/components/schemas/EquityListItem"
        - type: object
          properties:
            volume:
              type: integer
              format: int64
              nullable: true
            high:
              type: number
              format: double
              nullable: true
            low:
              type: number
              format: double
              nullable: true
            open:
              type: number
              format: double
              nullable: true
              description: Session reference open (daily bar) when available
            exchange:
              type: string
              example: EGX
            tvSymbol:
              type: string
              description: Upstream chart symbol identifier (non-secret)
            profile:
              nullable: true
              allOf:
                - $ref: "#/components/schemas/EquityScannerProfile"
            relatedStocks:
              type: array
              items:
                $ref: "#/components/schemas/EquityRelatedStockRow"
    EquityScannerProfile:
      type: object
      description: TradingView Egypt scanner snapshot (indicative; may differ from official filings).
      properties:
        tickerName: { type: string }
        description: { type: string, nullable: true }
        close: { type: number, format: double, nullable: true }
        changePct: { type: number, format: double, nullable: true }
        volume: { type: integer, format: int64, nullable: true }
        marketCap: { type: number, format: double, nullable: true }
        dividendYield: { type: number, format: double, nullable: true }
        peTtm: { type: number, format: double, nullable: true }
        epsTtm: { type: number, format: double, nullable: true }
        revenueFyHistory:
          type: array
          items: { type: number, format: double }
          nullable: true
        sector: { type: string, nullable: true }
        industry: { type: string, nullable: true }
        floatShares: { type: integer, format: int64, nullable: true }
        sharesOutstanding: { type: integer, format: int64, nullable: true }
        beta1Y: { type: number, format: double, nullable: true }
        employees: { type: integer, nullable: true }
        isin: { type: string, nullable: true }
        currency: { type: string, nullable: true }
        netIncomeFyHistory:
          type: array
          items: { type: number, format: double }
          nullable: true
        recommendAll:
          type: number
          format: double
          nullable: true
          description: Aggregate technical rating −1…1 (TradingView Recommend.All)
        recommendMovingAverages:
          type: number
          format: double
          nullable: true
          description: Moving-average aggregate (TradingView Recommend.MA)
        recommendOscillators:
          type: number
          format: double
          nullable: true
          description: Oscillator aggregate (TradingView Recommend.Other)
    EquityRelatedStockRow:
      type: object
      required: [symbol, name, close, changePct]
      properties:
        symbol: { type: string }
        name: { type: string }
        close: { type: number, format: double }
        changePct: { type: number, format: double }
    HistorySeries:
      type: object
      required: [symbol, resolution, points]
      properties:
        symbol:
          type: string
        resolution:
          type: string
          example: 1d
        points:
          type: array
          items:
            type: object
            required: [t, o, h, l, c]
            properties:
              t:
                type: string
                format: date-time
              o:
                type: number
              h:
                type: number
              l:
                type: number
              c:
                type: number
              v:
                type: integer
                format: int64
                nullable: true
...
    JournalCreateRequest:
      type: object
      required: [instrumentId, side, quantity, price, executedAt]
      properties:
        instrumentId:
          type: string
          format: uuid
        side:
          type: string
          enum: [buy, sell]
        quantity:
          type: number
          exclusiveMinimum: 0
        price:
          type: number
          minimum: 0
        executedAt:
          type: string
          format: date-time
        note:
          type: string
          nullable: true
    JournalPatchRequest:
      type: object
      minProperties: 1
      properties:
        note:
          type: string
          nullable: true
        price:
          type: number
          minimum: 0
        executedAt:
          type: string
          format: date-time
    JournalEntry:
      type: object
      required:
        [
          id,
          instrumentId,
          code,
          displayNameEn,
          side,
          quantity,
          price,
          executedAt,
          createdAt,
        ]
      properties:
        id:
          type: string
          format: uuid
        instrumentId:
          type: string
          format: uuid
        code:
          type: string
        displayNameEn:
          type: string
        side:
          type: string
          enum: [buy, sell]
        quantity:
          type: string
        price:
          type: string
        executedAt:
          type: string
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    JournalListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/JournalEntry"
    PortfolioSummary:
      type: object
      required: [disclaimer, positions]
      properties:
        disclaimer:
          type: string
        positions:
          type: array
          items:
            type: object
            required: [instrumentId, code, displayNameEn, netQuantity]
            properties:
              instrumentId:
                type: string
                format: uuid
              code:
                type: string
              displayNameEn:
                type: string
              netQuantity:
                type: number
              averageCost:
                type: number
                nullable: true
              costBasis:
                type: number
                nullable: true
              lastPrice:
                type: number
                nullable: true
              marketValue:
                type: number
                nullable: true
              unrealizedPl:
                type: number
                nullable: true
              unrealizedPlPct:
                type: number
                nullable: true
              quoteAsOf:
                type: string
                format: date-time
                nullable: true
...